Fixed MSVC 08 project
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4749 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CClient_Game.cpp
@@ -1304,7 +1304,7 @@ void CClient::ProcessShot_Beam(shoot_t *shot)
 					if((p - (*w2)->getPos()).GetLength2() < wormsize*wormsize) {
 						int damage = wpn->Bm.PlyDamage;
 						if(wpn->Bm.DistributeDamageOverWidth) { damage /= width; if(damage == 0) damage = SIGN(wpn->Bm.PlyDamage); }
-						InjureWorm(*w2, damage, shot->nWormID);
+						InjureWorm(*w2, (float)damage, shot->nWormID);
 
 						goodWidthParts[j] = false;
 						worms.erase(w2);
